requirements.txt
openai==1.52.0
anthropic==0.36.2

jupyter==1.1.1
jupyterlab==4.2.5

pytest==8.3.3
python-dotenv==1.0.1

nltk==3.9.1
rouge-score==0.1.2
sentence-transformers==3.2.0
scikit-learn==1.5.2
